+++ cgogn/core/cmap/cmap2.h
@@ -37,9 +37,9 @@ class CMap2_T : public CMap1_T<MAP_TRAITS, MAP_TYPE>
 {
 public:
 
-	static const int32 DIMENSION = 2;
+	static const uint8 DIMENSION = 2;
 
-	static const int32 PRIM_SIZE = 1;
+	static const uint8 PRIM_SIZE = 1;
 
 	using MapTraits = MAP_TRAITS;
 	using MapType = MAP_TYPE;
